updated to send raw ex, removed resulting circ dep
