+++ core/dbt/exceptions.py
@@ -2,7 +2,7 @@ import builtins
 import functools
 from typing import NoReturn, Optional, Mapping, Any
 from dbt.events.functions import fire_event
-from dbt.events.types import GeneralWarning
+from dbt.events.types import GeneralWarningMsg, GeneralWarningException
 from dbt.node_types import NodeType
 from dbt import flags
 from dbt.ui import line_wrap_message, warning_tag
@@ -976,14 +976,14 @@ def warn_or_error(msg, node=None, log_fmt=None):
     if flags.WARN_ERROR:
         raise_compiler_error(msg, node)
     else:
-        fire_event(GeneralWarning(msg=msg, log_fmt=log_fmt))
+        fire_event(GeneralWarningMsg(msg=msg, log_fmt=log_fmt))
 
 
 def warn_or_raise(exc, log_fmt=None):
     if flags.WARN_ERROR:
         raise exc
     else:
-        fire_event(GeneralWarning(msg=str(exc), log_fmt=log_fmt))
+        fire_event(GeneralWarningException(exc=exc, log_fmt=log_fmt))
 
 
 def warn(msg, node=None):
