+++ core/dbt/task/test.py
@@ -1,7 +1,8 @@
 from distutils.util import strtobool
 
 from dataclasses import dataclass
-from dbt import utils
+from dbt.utils import _coerce_decimal
+from dbt.events.format import pluralize
 from dbt.dataclass_schema import dbtClassMixin
 import threading
 from typing import Union
@@ -170,7 +171,7 @@ class TestRunner(CompileRunner):
         test_result_dct: PrimitiveDict = dict(
             zip(
                 [column_name.lower() for column_name in table.column_names],
-                map(utils._coerce_decimal, table.rows[0])
+                map(_coerce_decimal, table.rows[0])
             )
         )
         TestResultData.validate(test_result_dct)
@@ -181,7 +182,7 @@ class TestRunner(CompileRunner):
 
         severity = test.config.severity.upper()
         thread_id = threading.current_thread().name
-        num_errors = utils.pluralize(result.failures, 'result')
+        num_errors = pluralize(result.failures, 'result')
         status = None
         message = None
         failures = 0
